issues 36, 45
git-svn-id: http://co19.googlecode.com/svn/trunk/co19/tests/co19@44 2ecb45c1-dfb3-dc95-330a-13f216d23a04

+++ src/LibTest/core/int/operator_div/int_operator_div_A01_t01.dart
@@ -17,7 +17,7 @@
 // @3rdparty sputnik-v1/lib/math_precision.js
 double getPrecision(num number) {
   var log2num = Math.log(number.abs()) / Math.LN2;
-  int pernum = log2num.ceil();
+  double pernum = log2num.ceil();
   return(Math.pow(2, -51 + pernum));
 }
